--- a/mixer/Cargo.toml
+++ b/mixer/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "mixer"
+version = "0.1.0"
+authors = ["Chun-Min Chang <chun.m.chang@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bitflags = "1.2"