Create an oxidized cubeb-mixer, which is adapted from [cubeb_mixer.cpp][0], as the new mixer module. The new mixer is able to mix from any input channels to any output channels, without predefined channel order.

The mixer would mix audio input data to audio output data based on the given input and output audio _channel layout_. There are at least 2 different concepts when refering to audio _channel layout_.
1. _channel count_: number of channels
2. _channel order_: the order of the audio channel

In the past, we have the following issues when mixing the audio
1. The _channel count_ and the _channel order_ of a _channel layout_ are predefined. In our current implementation, for a _stereo_ layout, its _channel count_ must be _2_ and its _channel order_ must be _[front-left, front-right]_. However, in macOS, users are allowed to configure an _N_-channel device to any _1-to-N_ channel device. For example, a _6_-channel output device can be configured to a _stereo_ device that only uses _2_ channels and the left _4_ channels should be _ignored/unused_. That is, a _stereo channel layout_ whose _channel count_ could be _6_ and its _channel order_ could be _[unused, front-right, unused, unused, front-left, unused]_.
2. If the _channel count_ of a _channel layout_ is different from the predefined one, the _channel layout_ would be marked as an invalid layout and the audio data would be [passed through from input buffer to output buffer][1], without redirecting the audio data to its mapping channel. For example, the data of indices _0_ and _1_ in the audio input buffer whose layout is _stereo: [front-left, front-right]_ should be copied as the data of indices _3_ and _1_  in the audio output buffer whose layout _stereo: [unused, front-right, unused, front-left]_.
3. If the _channel count_ of a _channel layout_ is the same as the predefined one, the current mixer would ignore the user-defined _channel order_ and mix the audio data by the predefined channel order. For example, a _stereo_ output with 2 channels would always be treated as its channel order is _[front-left, front-right]_. However, users are allowed to set _stereo_ output to _[front-right, front-left]_ in macOS.

1. We would ignore the user-defined channel layout and force the output device to apply our pre-defined channel layout, which is a standard _SMPTE_ layout
    https://github.com/ChunMinChang/cubeb-coreaudio-rs/blob/cd954ca94baafc1bd0c4c661f07b816e9263ed1b/src/backend/mod.rs#L2722-L2731

The new mixer is created to fix the above 1-3 problems. The core idea is to think the mixing as matrix multiplication. The input and output audio data, _in_audio_, _out_audio_, are _Mx1_, _Nx1_ matrix and the mixing coefficients are a _NxM_ matrix _m_ such that _out_audio = m x in_audio_

```
data in channel #1 ▸ │ Silence    │   │ 0, 0, 0, 0 │   │ FrontRight   │ ◂ data in channel #1
data in channel #2 ▸ │ FrontRight │ = │ 1, C, 0, L │ x │ FrontCenter  │ ◂ data in channel #2
data in channel #3 ▸ │ FrontLeft  │   │ 0, C, 1, L │   │ FrontLeft    │ ◂ data in channel #3
                            ▴                 ▴        │ LowFrequency │ ◂ data in channel #4
                            ┊                 ┊                ▴
                            ┊                 ┊                ┊
                        out_audio      mixing matrix m       in_audio

where C, L are coefficients for mixing FrontCenter and LowFrequency data 
to FrontLeft or FrontRight channels. The Silence channel is a unused channel in the output device, so its channel data would be zero.
```

The mixing coefficient matrix _m_ would be generated dynamically, based on given _in_audio_ and _out_audio_. As a result, the restriction of the channel order would be loosen and audio is able to be mixed from any channel layout to any channel layout, without any pre-defined channel order and limitation of channels' number.

[0]: https://github.com/kinetiknz/cubeb/blob/81432070256d4e536cfcac14fe2b90f13866f369/src/cubeb_mixer.cpp
[1]: https://github.com/kinetiknz/cubeb/blob/81432070256d4e536cfcac14fe2b90f13866f369/src/cubeb_mixer.cpp#L570-L584