@ChunMinChang 2019-11-04
I would probably make this new mixer as an independent crate so we could reuse it in Windows. But I'll do in another pull request.