@ChunMinChang 2019-11-19
> This should generally works. I'm interested in two things:
> 
>     * Do the inner loop optimizes well? There are function calls, etc. but the optimizer should see through that. I'm hoping it can succeed in removing the bound checks, but I'm not sure since this is using array indexing.

I dumped the MIR by `cargo rustc --lib -- --emit mir` but I barely recognize it. I'll use some benchmark crate like _criterion_ to does some tests. We used to have the special handling for some (common) channel count: https://github.com/kinetiknz/cubeb/blob/81432070256d4e536cfcac14fe2b90f13866f369/src/cubeb_mixer.cpp#L475. I'll try to use special handling to see if the special handling really matters, by the third-party benchmark crate.


>     * What do you think is the common case? To me, a matrixing multichannel mixer is more of an application concern and not really an audio IO concern. You take the file or the application, and its output channel count and mapping, then you have a look at what the audio system can do. Then if there are mismatch, we can rematrix.
> 
> 
> For example, a 5.1 file to a 5.1 system should be pass-through, using the OS facility to do that, but a 5.1 to a 2.0 system should be rematrixed. A 2.0 file on a 5.1 system should be trying to output stereo via the OS, and not rematrix.

I guess the most common case is both input and output are stereos (but I never saw the telemetry). In that case, we won't create the mixer. We will only create a mixer when the input channel layout is different from the output channel layout. Is this the problem you worry about?
https://github.com/ChunMinChang/cubeb-coreaudio-rs/blob/0920240e4166d2b562840c8062e149d63f7c3a02/src/backend/mod.rs#L2733-L2754