@padenot 2019-11-19
> We will only create a mixer when the input channel layout is different from the output channel layout. Is this the problem you worry about?

We should try to not create the mixer if the input is stereo and the output is more than stereo, and instead try to open a stereo stream on AudioUnit, that will decide on the mappping. This is to solve bugs like bug 1468381.