@ChunMinChang 2019-11-19
> I'll use some benchmark crate like _criterion_ to does some tests. We used to have the special handling for some (common) channel count: https://github.com/kinetiknz/cubeb/blob/81432070256d4e536cfcac14fe2b90f13866f369/src/cubeb_mixer.cpp#L475. I'll try to use special handling to see if the special handling really matters, by the third-party benchmark crate.

I've introduced the _criterion_ crate for the benchmark. Now, the benchmark test could be run by `cargo bench` (There would be a plot report at _target/criterion/report/index.html_ if _gnuplot_ is installed).

The code with special handling was compared with the original implementation in 2 scenarios
- downmix from 5.1 to stereo
- upmix from mono to stereo

In short, with the special handling (code below), the `mix` would be slightly faster. When downmixing the audio 5.1 to the stereo with `512` _frames_, the special handling would be faster about _1 us_.

But I am not sure if it's worth adding complexity to gain this improvement. What do you think?

## Benchmark result

### Without special handling
---------
downmix_f32             time:   [19.276 us 19.345 us 19.419 us]                         
                        change: [+0.0700% +0.8003% +1.6345%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

downmix_i16             time:   [19.263 us 19.507 us 19.780 us]                         
                        change: [-1.8844% -0.2238% +1.5090%] (p = 0.79 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

upmix_f32               time:   [8.2549 us 8.2852 us 8.3193 us]                       
                        change: [-1.8970% -1.0369% -0.1848%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

upmix_i16               time:   [9.2803 us 9.3130 us 9.3497 us]                       
                        change: [-0.1100% +0.8325% +1.6188%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

### With special handling
-------------
downmix_f32             time:   [18.069 us 18.184 us 18.301 us]                         
                        change: [-0.2277% +0.4292% +1.0855%] (p = 0.21 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

downmix_i16             time:   [18.166 us 18.291 us 18.421 us]                         
                        change: [-1.2536% -0.3453% +0.5522%] (p = 0.45 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

upmix_f32               time:   [8.1000 us 8.1291 us 8.1582 us]                       
                        change: [+0.2135% +0.9894% +1.7244%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

upmix_i16               time:   [9.0213 us 9.0667 us 9.1131 us]                       
                        change: [-0.9681% -0.3106% +0.3800%] (p = 0.36 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

## Code with special handling for common input channel count
```rust
#[derive(Debug)]
pub struct Mixer<T>
where
    T: Copy + Debug + MixingCoefficient,
    T::Coef: AddAssign
        + Copy
        + Debug
        + Default
        + Mul<T::Coef, Output = T::Coef>
        + Add<T::Coef, Output = T::Coef>,
{
    coefficient: Coefficient<T>,
    coef_indices: Vec<Vec<usize>>,
}

impl<T> Mixer<T>
    ....

    pub fn new(input_channels: Vec<Channel>, output_channels: Vec<Channel>) -> Self {
        let coefficient = Coefficient::create(input_channels, output_channels);
        let mut coef_indices = Vec::new();
        for (i, _) in coefficient.output_channels().iter().enumerate() {
            let mut indices = Vec::new();
            for (j, _) in coefficient.input_channels().iter().enumerate() {
                if coefficient.get(j, i) != T::Coef::default() {
                    indices.push(j);
                }
            }
            coef_indices.push(indices);
        }
        Self {
            coefficient,
            coef_indices,
        }
    }

    pub fn mix(&self, input_buffer: &[T], output_buffer: &mut [T]) {
        assert_eq!(
            input_buffer.len(),
            self.input_channels().len(),
            "input slice must have the same size as the input channel's one."
        );
        assert_eq!(
            output_buffer.len(),
            self.output_channels().len(),
            "output slice must have the same size as the output channel's one."
        );
        for (i, output) in output_buffer.iter_mut().enumerate() {
            // T must implement Default that returns a zero value from default().
            let mut value = T::Coef::default(); // Create a zero value.

            match self.coef_indices[i].len() {
                0 => {
                    // do nothing
                }
                1 => {
                    let idx = self.coef_indices[i][0];
                    value =
                        T::to_coefficient_value(input_buffer[idx]) * self.coefficient.get(idx, i);
                }
                2 => {
                    let idx1 = self.coef_indices[i][0];
                    let idx2 = self.coef_indices[i][1];
                    value = T::to_coefficient_value(input_buffer[idx1])
                        * self.coefficient.get(idx1, i)
                        + T::to_coefficient_value(input_buffer[idx2])
                            * self.coefficient.get(idx2, i);
                }
                n => {
                    for k in 0..n {
                        let idx = self.coef_indices[i][k];
                        value += self.coefficient.get(idx, i)
                            * T::to_coefficient_value(input_buffer[idx]);
                    }
                }
            }
            *output = T::from_coefficient_value(
                value,
                self.coefficient.would_overflow_from_coefficient_value(),
            );
        }
    }

    ....
```